+++ .eslintrc.js
@@ -6,5 +6,7 @@ module.exports = {
     'no-console': 'off',
     '@typescript-eslint/no-implied-eval': 'off',
     '@typescript-eslint/no-throw-literal': 'off',
+    'no-plusplus': ['error', { allowForLoopAfterthoughts: true }],
+    'no-param-reassign': ['error', { props: false }],
   },
 };
