+++ app/emitter.ts
@@ -1,0 +1,3 @@
+import { EventEmitter } from './helpers/EventEmitter';
+
+export const emitter = new EventEmitter();
