+++ app/helpers/EventEmitter.ts
@@ -3,8 +3,6 @@ type Callback = (data?: any) => void;
 export class EventEmitter {
   callbacks: Record<string, Callback[]> = {};
 
-  constructor() {}
-
   on(event: string, cb: Callback) {
     if (!this.callbacks[event]) this.callbacks[event] = [];
     this.callbacks[event].push(cb);
