+++ app/helpers/maxMagnitude.ts
@@ -1,13 +1,16 @@
-export function maxMagnitude(/* arguments */) {
+export function maxMagnitude(...array: number[]) {
   let maxIndex: null | number = null;
   let maxValue = -1;
   for (let i = 0; i < arguments.length; i++) {
-    const abs = Math.abs(arguments[i]);
+    const abs = Math.abs(array[i]);
     if (abs > maxValue) {
       maxIndex = i;
       maxValue = abs;
     }
   }
-  if (maxIndex == null) return null;
-  return arguments[maxIndex];
+
+  if (maxIndex == null) {
+    return null;
+  }
+  return array[maxIndex];
 }
