+++ app/helpers/objectMap.ts
@@ -1,4 +1,4 @@
-export function objectMap(object, mapFn) {
+export function objectMap(object: Record<string, any>, mapFn) {
   return Object.keys(object).reduce((result, key) => {
     const value = mapFn(object[key], key);
     if (value != null) {
